requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
pinecone-client
bitsandbytes
newspaper3k==0.2.8
pydantic